+++ src/config.h
@@ -248,6 +248,7 @@ typedef struct {
   int options_fontsize;
   string old_options;
   bool dim_margins;
+  bool status_line;
   bool old_xbuttons;
   // Legacy
   bool use_system_colours;